refactor: Edit import paths for tests and restructure option loop
